Add files via upload
Hotel management project designed in Java and linked to Oracle databases. Room management, customer check-in and check-out, analysis for each room and analysis for room management. Hall management, customer check-in and check-out, analysis for halls and parties, and for party and hall management. Analysis for each part of the project. Dashboard plus analysis with a pie chart and a vertical chart. Complete analysis of the project plus analysis with a pyramid chart and knowing the profits as a percentage of the previous month and printing each analysis.